Improved app architecture and added some animations
